moved swagger files and added schemas for validation
